20_10_17

+++ inc/framework.h
@@ -1,6 +1,6 @@
 #pragma once
-void window(int width, int height, bool fullscreenMode=false);
-void window(const char* caption, int width, int height, bool fullscreenMode=false);
+void window(float width, float height, bool fullscreenMode=false);
+void window(const char* caption, float width, float height, bool fullscreenMode=false);
 bool msgProc();
 #define program() void gmain()
 #define repeat() while(msgProc())
